using automata instead of stream processor
Using the stream processor it is very hard to efficiently implement
stateful functions, for instance mean and sma. This comes from the
fact that from the moment a value has been produced the state is
completely forgotten.

Using the automata on the other hand implementing those functions is
now much simpler as it always returns a new version of itself. The
only drawback is that now we must use types like Maybe to represent
delays.