unixpipe now does only atomic writes
Writes to an unix pipe are atomic when the size written is less than
PIPE_BUF. The code now never writes more than PIPE_BUF bytes to a
pipe, even though currently there is only one writer.

Another change is that we are using O_RDWR instead of O_WRONLY to open
for writing. This is undefined under Posix, but Linux allows the open
succeed even without any reader available (which would otherwise fail
or block). This ties yet more the code to linux servers, but currently
we have no plans for other posix platforms (statement of doom).