dmproc reads from a pipe which internally is a broadcast channel that procs threads can attach to
Previously the xmpp component was responsible for receiving the events
somehow and feeding them into the dmproc connection. This commit
changes this drastically by creating a broadcast channel into the
dmproc that procs threads can attach to.

Another big change is the use of regular expressions instead of
globs. As such, it breaks compatilibity with previous release, hence
the reason we are changing it to v2.

The new code also creates 2N less threads than the previous one.