improving development building/testing tools
The improvements on dev.make are that it now checks all required
programs writing everything into the $(userfile). The error messages
are much more user friendly too.

dmtry has been deleted, meaning that from this commit on we are no
longer shipping it.

Lastly, the following bugs have been fixed:

* discover.sh depends on dash (`echo' must interpolate \n properly)

* trylib.sh#run attempts to remove all socket files, pid files and
  kill running procs prior starting a new test. Before doing this a
  failing test could compromise one or more following tests;