initscripts allows changing the / filesystem. This allows using virtualenv, although changing a couple of variables is still required.
The /etc/init.d scripts uses a variable $FSROOT that allows using a
different root path. Note that this must be available prior invoking
the script:

   $ env FSROOT=/foo/bar /etc/init.d/leela-:SERVICE ...

This will load /foo/bar/etc/default/:SERVICE instead of
/etc/default/:SERVICE. You will also need to change a couple of
variables for this to work, but it should be possible without changing
the init script itself.