implementing /v1/past24 using new twisted http service
Many improvements on LeelaWebHandler, most important being the
addition of the storage object into the http_protocol and disabling
xsfr tags and keep_alive.

Debug information that is available on the response is now only
enabled upon explicit requested (debug=true on query string). It now
contains more information, namely, the time taken to process the
request and the stack trace in case of error.

This commit also introduces an catch-all handler which always returns
404. This is now used when all other routes fail, to make sure we
always generate a json response.