deprecating /v1/[data/]range/:key url in favor of /v1/[data/]:key
Both resources will continue to work, but adding a note stating the
former is deprecated in favor of the later.

The try/golden directory has been renamed to try/smoke. This will make
it easier to understand, I presume, as the name is far more popular.

bugfix: That was odd: connecting twice to the databus. One should take
        a lot more care reviewing its own code to avoid writing such
        nonsense.