Implementing a O(1) space SMA algorithm (previously O(N))
Previously we had a window that shifted each element by one, as follows:

  . . . . . . . . . . . . . .
    |          |
    +----------+
  ---- move(1) -->

In practice, we had to maintain a buffer with exactly N elements.

The new algorithm now maintains two windows. Instead of shifting each
element we replace the old window with the new one when the later gets
full.

This makes the window size variable. It ranges from N to 2N, but that
seems not be a problem for our data.