implementing sma in terms of ewma
Using EWMA to implement SMA seems to be a better solution than using
an window size that varies over time [which probably doesn't have a
name :-)]. Both models are very similar and EWMA is a very efficient,
space constant, function to implement.

The implementation uses $1-2/(n+1)$ as the ⍺ value. This should give
us a reasonable approximate to the SMA function.

The implementation also takes care of warming up the EWMA. It waits
$(n+1)/2$ data points to arrive before finally switching over to EWMA.