Implementing couter, derive and absolute metric engine
The derive and counter follows closely the RRD implementation. We
don't accept data in the past and flush it only when new data is
received.

The absolute is able to sum up events that are arriving too fast. Thig
might be very useful to use it as an API, to measure requests per
second from different web servers.