implementing the timeline server
The serves exposes two unix sockets. The first one is the input, which
the metrics will arrive. The second is the broadcast socket, which
interested parties should attach themselves the receive events.

The implementation maintains an internal clock of 60 seconds (so
events will not be emitted faster than this [gauge is the sole
exception].

Currently we are forking 1 worker thread to deal with metrics. We can
increase this number as much as we like, but there will be a single
thread per wire, although currently it is possible to assign multiple
threads to work on the same queue.

The databus degrades linearly with the number of threads, so it is
advisable to not increase much this number.