removing support for tpq, there is no sense to fork multiple threads per queue
The implementation no longer makes use of STM for the Timeline. We
have to lock over the entire queue, so adding multiple threads will
probably make things worse.

The major bootleneck right now is that Databus is linear over the
queue size. However, as all frontends are in python, this is not a
concern, as this code is much faster than any python parser we have
comeup with.