new graph schema, a better approach toward supporting supernodes
In the previous schema, it was necessary to load the entire node to
filter a label. So, nodes with many edges be really hard to handle.

The new storage interface allows the query to load edges by label,
thus reducing by a great deal the number of nodes to read, plus only
reading the nodes necessary to perform the query.

Considering the defaults, in the previous model we would limit any
node to 10k edges. Now, a node may have 10k labels and each label may
have 10k nodes (edges), improving the storage by a 10k factor.

It is important to notice these limits applies only on the search. It
is currently possible to store many more edges per node.