warpdrive: rewriting the network protocol to adhere the documentation [leela-network-iterator] and using the new storage protocol
This changes effectively removes the limits on the graph. We can now
in theory read graphs of any size, including supernodes, without
degrading performance too much.

Obviously it may take a long time to consume everything, but the
performance should be the same independently of the size of the graph.

This commit also drops the redis as a cache backend, as it is very
hard get cache coherency right.