match allows vertex anywhere in the query
Previously the compiler would reject this query:

  match (a) -[...]> (b)

As the only place where a node could be provided was in the head of
the path. This commit adds support for this queries.

No actual changes to the backend protocol have to be made. As blackbox
always use the >= comparator we simply limit the number of rows to 1,
which effectively implements the getEdge funcionality.