bugfix: path queries were terminating early if intermediate nodes had no labels
To reproduce this problem:

  make (foo) -[bar]> (0)
  make (foo) -[bar]> (1)
  make (foo) -[bar]> (2) -[foo]> (baz)
  make (foo) -[bar]> (3)

Depending on the order on the database, the following query may fail
to produce a value:

  path (foo) -[bar]> () -[foo]> ()

The fetch function issues an EOF value when the backend reaches its
end. The way it is implemented, the fetch function must issue at least
one Chunk value. Failing to do so aborts the cursor.

We've changed the backend to inform the current page it is producing,
thus:

  Right (0, []) -> Chunk
  Right (_, []) -> EOF