implement thet stat statement which hides the zookeeper cluster from the clients
The motivation is to make the client library simpler: we no longer
need to depend on zookeeper libraries.

As each warpdrive instance keeps the zookeeper state in memory, we
simply expose this data through the zmq protocol.

The naming module, thus, simply issue an stat to any single machine in
order to make updates to the current perceived cluster.