make (a) now creates a guid that is no longer deterministic
Previously we were creating a guid using a cryptographic hash function:

  hash(namespace + name)

This makes it unnecessary to maintain a `naming' table as always know
what the hash will be but there were several drawbacks.

The first issue was that the protocol didn't enforce users creating
the node prior creating any edges, which is required in order to
resolve guids into names.

The second issue is related to deletes. In short, the delete is not
atomic and we need to make sure no user would re-create the node while
the delete is running.

Thus, using a different guid everytime a node is created makes it
easier to overcome these problems:

  1. we are changing the syntax to only use guids, forcing the user to
  register the node prior using it;

  2. deletes no longer require an lock as that guid will be never used
  again. Blackbox is now responsible for creating the guids.