blackbox now uses uuid to uniquely identify nodes
This commit makes blackbox responsably for registering nodes. GUIDs
are no longer a pure function (i.e. they are randomly generated).

To achieve this, we use the new cassandra lightweight transactions so
that `putname' message returns the GUID assigned to that name.

This have many implications the most important being a single
partition must be able to hold all GUIDs.