backend no longer uses a qdevice and allows user to provide page + limit
This is a huge optimization. Previously we were using a monad to
execute the query and, in that version, everything was serial. Many
theads were forked but there was no concurrency.

This new version uses a pipeline instead. The Comp module now returns
a list with all the paths that we need to visit and we visit them in
parallel.