giving up of cassandra lightweight transactions
We are getting way too much write timeouts to the point we can't use
it. This is happening with fairly low loads.

For now we are falling back to quorum, hoping this will give us strong
consistency. and are now vulnerable to race conditions, but the way we
use it, races are rare enough to go ahead with this.