Pool#use function allows using resources safely
Pool has being used only to create and destroy resources. No clients
ever nedded to make use of any of the availble resouces.

The upcoming redis storage implementation demands to be able to use
the resources, which in this case are connections with the redis
server.

This function allows user to safely use pool resources while other
threads might be updating the pool (creating or destroying).