adding classes to vertex nodes
Each vertex have a class, providing a type for vertex objects. In the
future users may query vertexes based on their classes.

Another use for this is to index attributes. We will use the class
information as the (unique) index name.

The syntax has not been changed, but the vertex names must adhere to
the following syntax:

    make (foo::bar)

The string before the :: is the class.