using init callback to connectd to leela
This is a bugfix. Collectd initialization phase is, roughly, as
follows:

    parse_config
    daemonize
    init_plugins
    ...

*leela_lql_cursor_init* starts a thread to monitor the leela
cluster. In the previous version, to avoid the use of a global
variable, we were initializing the leela context after parsing the
config. After forking, the monitor thread was no longer available.

The issue did not happened with collectdmon because no forking was
performed.