no longer using hostname as prefix; user may use a new "prefix" config option
The reason is to make easier to read attributes from many nodes. In
the previous layout, each metric was unique since the hostname was the
prefix. The new default make different nodes share the metric name.

To continue using the old default, one must use the new prefix cfg
option, using the hostname as its value:

    Prefix "<hostname>/"