enabling MAC-based authentication
This patch introduces a file-based database that must be managed
externally. It continuously watches this file for changes (which
should be atomic) and reloads its setting.

In the future we want to make this part of the leela system itself,
implementing some kind of GRANT command.

Also, the naming module now allows users do define the username &
password to use, using an environment variable.