use redis as cache for the last t-attr value
We store the last value of each t-attr on redis, with a ttl of 86400
seconds. This is used to reduce write load on t-index table.

In the future we may let users to access the index as well using it to
implement more data types that depends on previous values.