add attr last statement
The syntax is as follows:

  attr last <guid | *> "<attr | glob>";

It retrieves all attributes that match a given glob pattern on a
guiven uid or all uids when * is provided.

Noticed that this uses the cache and there are no guarantees about the
results. In special it may return no results at all, duplicate
results, or fewer results than actually exists.