allow user of dealer interface to cancel requests and tune zmq queues [hwm]
This patch enqueues an mvar and returns a function that clears the
contents of this mvar allowing users to cancel requests in case they
have not been sent already [and doing nothing otherwise].

Prior this patch, hwm was defined to (0, 0) which means unbounded
queues. After the first timeout, every single subsequent request also
times out since there was no way to remove these expired requests from
the zmq queue.

The new value for hwm is (1000, 1), which means we are using solely
our own queue for buffering requests. This allows the server to
recover from spikes, as timeouts will probably get removed from the
queue.