new timeout manager and refactor ioloop
The first change is about performance whereas the second one is to
improve the code base.

The previous timeout mechanism used haskell thread [like
System.Timeout does]. This is elegant and simple, but does not scale
well. To put this in perspective, a request that writes 320 t-attr
values forks about 11 threads just for timeout [1 for the request + 10
for each backend call]. This is a waste.

Inspired by the warp http server [1], the new timeout manager uses a
linked list of IORefs. The implementation follows the ideas of the
warp server and it has worked really well.

The second change is done to simplify the code base. We now use a MVar
for incoming request instead of TBQueue. This makes the code alternate
between receiving and processing requests.

The new ioloop also saves one thread. In the past, two threads were
monitoring read and write events and the other thread handling the
requests. Now the ioloop performs the read and write concurrently,
each thread polling for events. This makes the code a lot simpler.

Lastly, significant changes have been made to WarpServer and QDevice
modules, also for performance and maintainability sake.

[1] http://aosabook.org/en/posa/warp.html