lql_cursor_restart gives the ability to reuse a cursor
The motivation for this are queries that produces many subqueries. If
a machine is failing this will probably fail the entire request there
is a high probability that using multiple cursors will reach all
available machines.

With this addition, the user is able to keep reusing the same cursor
in order to reduce the error propagation.