tracks number of active connections and defines a soft and hard limit
The timeout module now tracks the number of active handles which is
the number of active connections.

The server defines a soft limit and a hard limit for the number of
active connections. These limits are defined based on the number of
capabilities. The soft limit is used to introduce a delay before
scheduling the request whereas the hard limit drops the request
silently.