sendMsg can no longer fail
We now use an unbounded channel that supports cancelation. In other
words we accept every message but the requester can cancel the message
before it gets sent [with no guarantees though].

This has been used for some time but the interface was still the old
one, that used to use bounded queues.

This patch finally updates the sendMsg interface.