use 12 t-attr colum families
We used to use a single t-attr table. This table grows very fast under
high loads and make it hard to scale cassandra in a timely manner.

Using 12 tables makes it easy to compaction to catch up, which is
important since we are using leveled compaction strategy.

Archiving is also much easier since for most of the time the column
families will be essentially read-only.