add warpgrep to the project
warpgrep is a new feature that allows users to monitor live
events. Using this it is possible to monitor graph changes as well as
attribute changes. It works by forking write messages that goes to the
storage system so that users can monitor these events in real-time. It
works roughly as follows:

    +-----------+
    |           |
    | warpdrive |----+------> storage
    |           |    |
    +-----------+    |
                     +------> warpgrep

The message above is not a precise description about the message
flow. In this implementation, the message only goes to the warpgrep
after the storage has acknowledged the write.

This commit does not yet expose this feature via LQL.