warpgrep and msgpack
This commit got just too long, anyway here we go.

In this commit we add the syntax and support for the message bus
[hereafter named grep]. This allows users registering queries using the
following syntax:

    grep attr <GUID|*> "REGEX"          -- monitor k-attr
    grep attr <GUID|*> "REGEX" []       -- t-attr

    grep make "REGEX" "REGEX"           -- make vertex
    grep make <GUID|*> "REGEX" <GUID|*> -- make link

    grep kill <GUID|*>                  -- kill vertex
    grep kill <GUID|*> "REGEX" <GUID|*> -- kill link

This register the queries on all available warpgrep at the time and must
be done regularly as it expires after a while. After doing this you must
connect manually to the warpgrep instances [which you can discover using
the stat command] though this will done in the library in the future.

The second commit changes the protocol between warpdrive and blackbox
instances. It uses msgpack instead of zmq frames for performance
reasons.