ignore fetch requests that can't possibly exists
We've seen FETCH requests from broken clients that is in some sort of
infinite loop. This generates some inconvenient load on the server.

The implementation makes use of the `useTimeout' to mark checkpoints
over time. When the current time exceeds the checkpoint time by
'useTimeout' seconds, we store that new information. Since FH numbers
are always increasing, this denotes the minimum FH number that could
possibly exist in the server before the timeout kicks in.

When an handle does not exist on the server we compare it against the
last checkpoint value. If the FH is less than the checkpoint'ed FH we
silently drop the requests.