create basic centos e debian docker images
The ``automation/docker/makeimg.sh`` script creates docker images with
the bare minimum [as defined by debootstrat and rinse] packages required
to compile zeromq.

Incidentally, this is also the leela starting point. From there it is
easy to adapt the container to include support for other components and
to build packages.