leela high-level protocol definition
We now have a protocol describing the request-response model. This is
going to change everything.

The most sensible change is the requirement of stateless servers, which
is not the case today. The main motivation for this is the ability to
resume an request from a given point on any instance.

This is very important. It allows us to do maintenance without
disrupting the service, implement efficient retry logic and distribute
load more evenly amongst the nodes.