test_cassandra_metrics using the trigger
This code is using the cassandra trigger to index metrics data. The
schema also register the trigger it that is not already registered.

Also, a small number of changes have been made to the trigger code. In
particular, we're not using the sequence-index because inserts on data
table are using transactions and the index has a different partition
key. For now we have no reliable solution to index data on this table.