use secondary index to scan sequence table
The reason for this change is to maintain the secondary indexes
consistent.

Consider the previous schema:

    create table sequence
    ( plane bigint
    , seqid bigint
    , object blob
    , primary key ((plane, object))
    );

    create table sequence_idx
    ( plane bigint
    , seqid bigint
    , object blob
    , primary key (plane, seqid)
    );

Since we use C* Lightweight Transactions, there is no way to use
triggers to populate the sequence_idx table so the only solution is to
do this on the client code, which may lead to serious consistency
problems.

However, the main purpose of this index is to scan a given `plane` in
order to cache its contents. Thus, this patch changes the schema to
this.

    create table sequence
    ( plane bigint
    , seqid bigint
    , object blob
    , primary key (object, plane)
    );

    create index on sequence (plane);

Which allows to scan a plane on its entirely and be sure this index is
consistent [as consistent as secondary indexes can get]:

    select *
    from sequence
    where plane = X and token(object) > token(O)
    limit N;